Improved CSV file reading, fixed typo in manual.
Added support for reading expression data CSV files where the format of the header row is such that it begins with a label for the gene ID column, followed by sample IDs after this. In the earlier version, no label should be used for the gene ID column. Otherwise, the program would fail to properly match samples to their correct gene expression values. Also updated the manual to fix a typo and an outdated reference to a CNA CSV input option. #2